requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
certifi==2025.8.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.8.0
django-debug-toolbar==6.0.0
django-filter==25.1
django-prometheus==2.4.1
django-use-email-as-username==1.4.0
django_dramatiq==0.14.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dramatiq==1.18.0
dramatiq-dashboard==0.4.0
MarkupSafe==2.0.1
packaging==25.0
prometheus_client==0.22.1
psycopg2-binary==2.9.10
pydantic==2.11.9
pydicom==3.0.1
redis==4.5.4
requests==2.32.5
safetensors==0.6.2
six==1.17.0
sqlparse==0.5.3
tifffile==2025.9.9
typing-inspection==0.4.1
typing_extensions==4.15.0
drf-yasg==1.21.10
pillow==11.3.0
pika==1.3.2